Move Layout to storm and add preloader
