+++ client/web/src/SourcegraphWebApp.tsx
@@ -33,7 +33,6 @@ import { authenticatedUser as authenticatedUserSubject, AuthenticatedUser, authe
 import { ComponentsComposer } from './components/ComponentsComposer'
 import { ErrorBoundary, RouteError } from './components/ErrorBoundary'
 import { FeatureFlagsProvider } from './featureFlags/FeatureFlagsProvider'
-import { Layout } from './Layout'
 import { LegacyRoute, LegacyRouteContextProvider } from './LegacyRouteContext'
 import { PageError } from './PageError'
 import { createPlatformContext } from './platform/context'
@@ -45,6 +44,8 @@ import { setQueryStateFromSettings, useNavbarQueryState } from './stores'
 import { AppShellInit } from './storm/app-shell-init'
 import { UserSessionStores } from './UserSessionStores'
 import { siteSubjectNoAdmin, viewerSubjectFromSettings } from './util/settings'
+import { Layout } from './storm/pages/LayoutPage/LayoutPage'
+import { loader } from './storm/pages/LayoutPage/LayoutPage.loader'
 
 export interface StaticSourcegraphWebAppContext {
     setSelectedSearchContextSpec: (spec: string) => void
@@ -260,6 +261,8 @@ export const SourcegraphWebApp: FC<SourcegraphWebAppProps> = props => {
         () =>
             createBrowserRouter([
                 {
+                    // The layout page is needed for every route so we do not need to lazy-load it.
+                    loader,
                     element: <LegacyRoute render={props => <Layout {...props} />} />,
                     children: props.routes,
                     errorElement: <RouteError />,
