+++ client/web/src/global/Notices.test.tsx
@@ -4,28 +4,8 @@ import { Notices } from './Notices'
 
 describe('Notices', () => {
     test('shows notices for location', () =>
-        expect(
-            renderWithBrandedContext(
-                <Notices
-                    location="home"
-                    settingsCascade={{
-                        subjects: [],
-                        final: {
-                            notices: [
-                                { message: 'a', location: 'home' },
-                                { message: 'a', location: 'home', dismissible: true },
-                                { message: 'b', location: 'top' },
-                            ],
-                        },
-                    }}
-                />
-            ).asFragment()
-        ).toMatchSnapshot())
+        expect(renderWithBrandedContext(<Notices location="home" />).asFragment()).toMatchSnapshot())
 
     test('no notices', () =>
-        expect(
-            renderWithBrandedContext(
-                <Notices location="home" settingsCascade={{ subjects: [], final: { notices: undefined } }} />
-            ).asFragment()
-        ).toMatchSnapshot())
+        expect(renderWithBrandedContext(<Notices location="home" />).asFragment()).toMatchSnapshot())
 })
